Trends title in the detailed page. Graphs margin right
