+++ app/src/components/Trends/TrendsTitle.css
@@ -1,0 +1,12 @@
+.TrendsTitle {
+  margin: 12px;
+  text-align: center;
+  border-bottom: 2px dashed #000;
+  cursor: pointer;
+  line-height: 1 !important;
+  margin-bottom: 20px !important;
+}
+
+.TrendsExplorePage .TrendsTitle {
+  display: inline-block;
+}
