+++ app/src/components/Trends/TrendsTitle.js
@@ -1,0 +1,35 @@
+import React from 'react'
+import { Popup } from 'semantic-ui-react'
+import { style } from '..//HelpPopup/HelpPopup'
+import './TrendsTitle.css'
+
+const TrendsPageTitleWithPopup = () => (
+  <Popup
+    trigger={<h1 className='TrendsTitle'>Social Trends</h1>}
+    position='bottom center'
+    on='hover'
+    style={style}
+    content={
+      <div>
+        <p>
+          Our Social Trends search is unique in the industry, based on channels,
+          sites and forums.
+        </p>
+        <p>
+          We focus only on channels where 90% of the discussions about crypto,
+          including "insider" sites not open to public Google search.
+        </p>
+        <h4>Combined results include:</h4>
+        <ul>
+          <li>Telegram channels</li>
+          <li>Subreddits</li>
+          <li>Discord Channels</li>
+          <li>BitcoinTalk</li>
+          <li>Other chats/forums</li>
+        </ul>
+      </div>
+    }
+  />
+)
+
+export default TrendsPageTitleWithPopup
